+++ forge/src/main/java/com/iamkaf/amber/mixin/LootTableAccessor.java
@@ -2,6 +2,8 @@ package com.iamkaf.amber.mixin;
 
 import net.minecraft.world.level.storage.loot.LootPool;
 import net.minecraft.world.level.storage.loot.LootTable;
+
+import java.util.List;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Mutable;
 import org.spongepowered.asm.mixin.gen.Accessor;
@@ -9,9 +11,9 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 @Mixin(LootTable.class)
 public interface LootTableAccessor {
     @Accessor("pools")
-    LootPool[] amber$getPools();
+    List<LootPool> amber$getPools();
 
     @Accessor("pools")
     @Mutable
-    void amber$setPools(LootPool[] pools);
+    void amber$setPools(List<LootPool> pools);
 }