+++ versions/1.19.1/fabric/src/main/java/com/iamkaf/amber/platform/FabricAmberEventHandlers.java
@@ -133,7 +133,8 @@ final class FabricAmberEventHandlers {
     static void registerClientCommandEvents() {
 
         ClientCommandRegistrationCallback.EVENT.register((dispatcher, registryAccess) -> {
-            CommandDispatcher<CommandSourceStack> commandsTemp = new CommandDispatcher<>();
+            CommandDispatcher<CommandSourceStack> commandsTemp =
+                    (CommandDispatcher<CommandSourceStack>) (CommandDispatcher<?>) dispatcher;
             ClientCommandEvents.EVENT.invoker().register(commandsTemp, registryAccess);
         });